+++ src/core/distributions/phylogenetics/tree/birthdeath/StateDependentSpeciationExtinctionProcess.h
@@ -50,8 +50,8 @@ namespace RevBayesCore {
         
         void                                                            drawJointConditionalAncestralStates(std::vector<size_t>& startStates, std::vector<size_t>& endStates);
         void                                                            recursivelyDrawJointConditionalAncestralStates(const TopologyNode &node, std::vector<size_t>& startStates, std::vector<size_t>& endStates);
-        void                                                            drawJointConditionalCharacterMap(std::vector<std::string*>& character_histories);
-        void                                                            recursivelyDrawJointConditionalCharacterMap(const TopologyNode &node, size_t start_state, std::vector<std::string*>& character_histories);
+        void                                                            drawStochasticCharacterMap(std::vector<std::string*>& character_histories);
+        void                                                            recursivelyDrawStochasticCharacterMap(const TopologyNode &node, size_t start_state, std::vector<std::string*>& character_histories);
         void                                                            numericallyIntegrateProcess(state_type &likelihoods, double begin_age, double end_age, bool use_backward, bool extinction_only) const; //!< Wrapper function for the ODE time stepper function.
         
     protected: