+++ src/core/math/distributions/DistributionInverseGamma.cpp
@@ -24,14 +24,15 @@ using namespace RevBayesCore;
  *
  * \brief InverseGamma probability density.
  * \param the shape parameter of the InverseGamma.
- * \param the scale parameter of the InverseGamma.
+ * \param the rate parameter of the InverseGamma.
  * \param x is the InverseGamma random variable.
  * \return Returns the probability density.
  * \throws Does not throw an error.
  */
-double RbStatistics::InverseGamma::pdf(double shape, double scale, double x) {
+double RbStatistics::InverseGamma::pdf(double shape, double rate, double x)
+{
     
-	return (pow(scale, shape) / RbMath::gamma(shape)) * pow(x, -(shape + 1.0)) * exp(-scale / x);
+	return (pow(rate, shape) / RbMath::gamma(shape)) * pow(x, -(shape + 1.0)) * exp(-rate / x);
 }
 
 
@@ -41,17 +42,18 @@ double RbStatistics::InverseGamma::pdf(double shape, double scale, double x) {
  *
  * \brief InverseGamma probability density.
  * \param the shape parameter of the InverseGamma.
- * \param the scale parameter of the InverseGamma.
+ * \param the rate parameter of the InverseGamma.
  * \param x is the InverseGamma random variable.
  * \return Returns the probability density.
  * \throws Does not throw an error.
  */
-double RbStatistics::InverseGamma::pdf(double shape, double scale, double x, bool isLog) {
+double RbStatistics::InverseGamma::pdf(double shape, double rate, double x, bool isLog)
+{
     
     //    double pr;
-    //    if (shape < 0 || scale <= 0) {
+    //    if (shape < 0 || rate <= 0) {
     //        std::ostringstream s;
-    //        s << "Cannot compute the pdf for the InverseGamma distribution for shape = " << shape << " and scale = " << scale;
+    //        s << "Cannot compute the pdf for the InverseGamma distribution for shape = " << shape << " and rate = " << rate;
     //        throw (RbException(s));
     //	    }
     // if (x < 0)
@@ -64,18 +66,18 @@ double RbStatistics::InverseGamma::pdf(double shape, double scale, double x, boo
     //	        if (shape < 1) return RbConstants::Double::inf;
     //	        if (shape > 1) return 0.0;
     //	        /* else */
-    //	        return isLog ? -log(scale) : 1 / scale;
+    //	        return isLog ? -log(rate) : 1 / rate;
     //	    }
     //
     // if (shape < 1) {
-    //	        pr = RbStatistics::Poisson::pdf(shape, x/scale, isLog);
+    //	        pr = RbStatistics::Poisson::pdf(shape, x/rate, isLog);
     //	        return isLog ?  pr + log(shape/x) : pr*shape/x;
     //	    }
     // /* else  shape >= 1 */
-    // pr = RbStatistics::Poisson::pdf(shape-1, x/scale, isLog);
-    // return isLog ? pr - log(scale) : pr/scale;
+    // pr = RbStatistics::Poisson::pdf(shape-1, x/rate, isLog);
+    // return isLog ? pr - log(rate) : pr/rate;
     
-    return isLog ? pdf(shape, scale, exp(x)) : pdf(shape, scale, x);
+    return isLog ? pdf(shape, rate, exp(x)) : pdf(shape, rate, x);
 }
 
 /*!
@@ -84,14 +86,15 @@ double RbStatistics::InverseGamma::pdf(double shape, double scale, double x, boo
  *
  * \brief Natural log of InverseGamma probability density.
  * \param the shape parameter of the InverseGamma.
- * \param the scale parameter of the InverseGamma.
+ * \param the rate parameter of the InverseGamma.
  * \param x is the InverseGamma random variable.
  * \return Returns the natural log of the probability density.
  * \throws Does not throw an error.
  */
-double RbStatistics::InverseGamma::lnPdf(double shape, double scale, double x) {
+double RbStatistics::InverseGamma::lnPdf(double shape, double rate, double x)
+{
     
-	return shape * log(scale) - RbMath::lnGamma(shape) + (-(shape + 1.0)) * log(x) - (scale/x);
+	return shape * log(rate) - RbMath::lnGamma(shape) + (-(shape + 1.0)) * log(x) - (rate/x);
 }
 
 /*!
@@ -100,17 +103,17 @@ double RbStatistics::InverseGamma::lnPdf(double shape, double scale, double x) {
  *
  * \brief InverseGamma cumulative probability.
  * \param the shape parameter of the InverseGamma.
- * \param the scale parameter of the InverseGamma.
+ * \param the rate parameter of the InverseGamma.
  * \param x is the InverseGamma random variable.
  * \return Returns the cumulative probability.
  * \throws Does not throw an error.
  */
-double RbStatistics::InverseGamma::cdf(double shape, double scale, double x) {
+double RbStatistics::InverseGamma::cdf(double shape, double rate, double x)
+{
 
-    double lowerIncompleteGamma = RbMath::incompleteGamma( scale/x, shape, RbMath::lnGamma(shape) );
-    double gamma = RbMath::gamma(shape);
+    double lower_incomplete_gamma = RbMath::incompleteGamma( rate/x, shape, RbMath::lnGamma(shape) );
     
-    return (gamma - lowerIncompleteGamma) / gamma;
+    return 1 - lower_incomplete_gamma;
 }
 
 /*!
@@ -119,24 +122,25 @@ double RbStatistics::InverseGamma::cdf(double shape, double scale, double x) {
  *
  * \brief InverseGamma quantile.
  * \param the shape parameter.inc
- * \param the scale parameter.
+ * \param the rate parameter.
  * \param p is the probability up to the quantile.
  * \return Returns the quantile.
  * \throws Does not throw an error.
  */
-double RbStatistics::InverseGamma::quantile(double shape, double scale, double p) {
-    
-    
+double RbStatistics::InverseGamma::quantile(double shape, double rate, double p)
+{
+    throw RbException("The quantile function of the inverse gamma distribution is not implemented yet!");
     // NOT IMPLEMENTED...
     return 0.0;
     
-	return RbStatistics::ChiSquare::quantile(p, 2.0 * shape) / (2.0 * scale);
-    //	return RbStatistics::Helper::pointChi2(p, 2.0 * shape) / (2.0 * scale);
+	return RbStatistics::ChiSquare::quantile(p, 2.0 * shape) / (2.0 * rate);
+    //	return RbStatistics::Helper::pointChi2(p, 2.0 * shape) / (2.0 * rate);
 }
 
 
-double RbStatistics::InverseGamma::rv(double shape, double scale, RandomNumberGenerator& rng) {
+double RbStatistics::InverseGamma::rv(double shape, double rate, RandomNumberGenerator& rng)
+{
     
-	return (RbStatistics::Helper::rndGamma(shape, rng) * scale);
+	return (RbStatistics::Helper::rndGamma(shape, rng) * rate);
 }