+++ src/revlanguage/datatypes/phylogenetics/characterdata/RlContinuousCharacterData.h
@@ -63,6 +63,10 @@ namespace RevLanguage {
         
         RevBayesCore::TypedDagNode<valueType>*                  dagNode;
 
+    private:
+
+        void                                                    initMethods(void);
+
     };
     
 }