+++ src/revlanguage/utils/GitVersion.cpp
@@ -1,4 +1,4 @@
 #include "GitVersion.h"
-const char *build_git_sha = "90e6ff";
-const char *build_date = "Mon Jan  2 15:56:50 CET 2017";
-const char *build_git_branch = "master";
+const char *build_git_sha = "db896d";
+const char *build_date = "Fri Mar  3 10:06:38 EST 2017";
+const char *build_git_branch = "development";