+++ src/revlanguage/workspace/RbRegister_Func.cpp
@@ -104,6 +104,7 @@
 #include "Func_checkNodeOrderConstraints.h"
 #include "Func_computeWeightedNodeOrderConstraintsScore.h"
 #include "Func_concatenate.h"
+#include "Func_concatenateFromVector.h"
 #include "Func_constructRootedTripletDistribution.h"
 #include "Func_extantTree.h"
 #include "Func_formatDiscreteCharacterData.h"
@@ -128,6 +129,7 @@
 /* Rate matrix functions (in folder "functions/phylogenetics/ratematrix") */
 #include "Func_blosum62.h"
 #include "Func_chromosomes.h"
+#include "Func_covarion.h"
 #include "Func_cpRev.h"
 #include "Func_dayhoff.h"
 #include "Func_DECRateMatrix.h"
@@ -278,6 +280,7 @@ void RevLanguage::Workspace::initializeFuncGlobalWorkspace(void)
         /* Rate matrix generator functions (in folder "functions/evolution/ratematrix") */
         addFunction( new Func_blosum62()                 );
         addFunction( new Func_chromosomes()              );
+        addFunction( new Func_covarion()                 );
         addFunction( new Func_cpRev()                    );
         addFunction( new Func_dayhoff()                  );
         addFunction( new Func_DECRateMatrix()            );
@@ -322,30 +325,31 @@ void RevLanguage::Workspace::initializeFuncGlobalWorkspace(void)
         addFunction( new Func_SampledCladogenesisRootFrequencies() );
 
 		/* Functions related to phylogenetic trees (in folder "functions/phylogenetics/tree") */
-        addFunction( new Func_branchScoreDistance()      );
-        addFunction( new Func_checkNodeOrderConstraints());
+        addFunction( new Func_branchScoreDistance()         );
+        addFunction( new Func_checkNodeOrderConstraints()   );
 //        addFunction( new Func_clade()                    );
         addFunction( new Func_computeWeightedNodeOrderConstraintsScore());
-        addFunction( new Func_concatenate()              );
+        addFunction( new Func_concatenate()                 );
+        addFunction( new Func_concatenateFromVector()       );
         addFunction( new Func_constructRootedTripletDistribution()            );
         addFunction( new Func_formatDiscreteCharacterData() );
-        addFunction( new Func_extantTree()                );
-        addFunction( new Func_maximumTree()              );
-        addFunction( new Func_mrcaIndex()                );
-        addFunction( new Func_nodeAgeByID()              );
+        addFunction( new Func_extantTree()                  );
+        addFunction( new Func_maximumTree()                 );
+        addFunction( new Func_mrcaIndex()                   );
+        addFunction( new Func_nodeAgeByID()                 );
         addFunction( new Func_phyloDiversity() );
-        addFunction( new Func_pomoStateConverter()       );
-        addFunction( new Func_pomoRootFrequencies()      );
-        addFunction( new Func_pruneTree()                );
-        addFunction( new Func_readPomoCountFile()       );
-        addFunction( new Func_simTree()                  );
-        addFunction( new Func_simCompleteTree()          );
-        addFunction( new Func_stitchTree()               );
-        addFunction( new Func_symmetricDifference()      );
-        addFunction( new Func_tmrca()                    );
-        addFunction( new Func_treePairwiseDistances()    );
-        addFunction( new Func_treeAssembly()             );
-        addFunction( new Func_treeScale()                );
+        addFunction( new Func_pomoStateConverter()          );
+        addFunction( new Func_pomoRootFrequencies()         );
+        addFunction( new Func_pruneTree()                   );
+        addFunction( new Func_readPomoCountFile()           );
+        addFunction( new Func_simTree()                     );
+        addFunction( new Func_simCompleteTree()             );
+        addFunction( new Func_stitchTree()                  );
+        addFunction( new Func_symmetricDifference()         );
+        addFunction( new Func_tmrca()                       );
+        addFunction( new Func_treePairwiseDistances()       );
+        addFunction( new Func_treeAssembly()                );
+        addFunction( new Func_treeScale()                   );
 
         /* Population genetics functions (in folder "functions/popgen") */
         addFunction( new Func_PattersonsD()      );