+++ CHANGELOG.md
@@ -6,6 +6,7 @@ This file contains a list of notable upgrades and features introduced to each re
 #### Added
 - `CHANGELOG.md` added
 - Support to provide autotuned acceptance rates for `mvScale`, `mvSlide`, and `mvSimplex` moves using the `tuneTarget` argument
+- Support to store simulated character states of internal nodes for `dnPhyloCTMC` and `dnPhyloCTMCClado` using the `storeInternalNodes` argument
 
 #### Changed